# tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AnimeTypes": {
				"type": "string",
				"enum": [
					"Special",
					"Movie",
					"OVA",
					"ONA",
					"TV",
					""
				]
			},
			"AnimeStatuses": {
				"type": "string",
				"enum": [
					"FINISHED",
					"UPCOMING",
					"UNKNOWN",
					"CURRENTLY",
					""
				]
			},
			"AnimeSeason": {
				"type": "string",
				"enum": [
					"SUMMER",
					"UNDEFINED",
					"WINTER",
					"SPRING",
					"FALL",
					""
				]
			},
			"AnimeTags": {
				"items": {
					"type": "string"
				},
				"type": "array"
			},
			"AnimeItem": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"sources": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"title": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/AnimeTypes"
					},
					"episodes": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/AnimeStatuses"
					},
					"animeSeason": {
						"properties": {
							"year": {
								"type": "number",
								"format": "double"
							},
							"season": {
								"$ref": "#/components/schemas/AnimeSeason"
							}
						},
						"required": [
							"year",
							"season"
						],
						"type": "object"
					},
					"unacceptable": {
						"type": "boolean"
					},
					"picture": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"synonyms": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"relations": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"tags": {
						"$ref": "#/components/schemas/AnimeTags"
					}
				},
				"required": [
					"_id",
					"sources",
					"title",
					"type",
					"episodes",
					"status",
					"animeSeason",
					"unacceptable",
					"picture",
					"thumbnail",
					"synonyms",
					"relations",
					"tags"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnimeList": {
				"items": {
					"$ref": "#/components/schemas/AnimeItem"
				},
				"type": "array"
			},
			"UserAnimeStatuses": {
				"type": "string",
				"enum": [
					"viewed",
					"abandoned",
					"postponed",
					"planned",
					"reviewing",
					"look"
				]
			},
			"UserAnimeStars": {
				"type": "string",
				"enum": [
					"0",
					"1",
					"2",
					"3",
					"4",
					"5",
					"6",
					"7",
					"8",
					"9",
					"10"
				]
			},
			"UserAnimeValues": {
				"properties": {
					"star": {
						"$ref": "#/components/schemas/UserAnimeStars"
					},
					"review": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/UserAnimeStatuses"
					}
				},
				"required": [
					"star",
					"review",
					"status"
				],
				"type": "object"
			},
			"UserAnime": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/UserAnimeValues"
				},
				"type": "object"
			},
			"AnimeDetail": {
				"$ref": "#/components/schemas/UserAnime"
			},
			"AnimeDetailList": {
				"items": {
					"$ref": "#/components/schemas/AnimeDetail"
				},
				"type": "array"
			},
			"UserSchema": {
				"properties": {
					"uid": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"photoURL": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"phoneNumber": {
						"type": "string",
						"nullable": true
					},
					"isAnonymous": {
						"type": "boolean"
					},
					"tenantId": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"uid",
					"displayName",
					"photoURL",
					"email",
					"emailVerified",
					"phoneNumber",
					"isAnonymous",
					"tenantId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorSchema": {
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			}
		}
	},
	"info": {
		"title": "backend-animeover",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/animedb": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"count": {
											"type": "number",
											"format": "double"
										},
										"animeList": {
											"$ref": "#/components/schemas/AnimeList"
										}
									},
									"required": [
										"count",
										"animeList"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"limit": "1",
											"tags": "[\"comedy\"]",
											"season": "WINTER",
											"page": "1"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "season",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/AnimeSeason"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/animedb/anime/{animeId}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnimeItem"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "animeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/animedb/tags": {
			"get": {
				"operationId": "GetAnimeTags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/animedetail/{animeId}": {
			"get": {
				"operationId": "GetOneDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnimeDetailList"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "animeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/animelist": {
			"get": {
				"operationId": "UserAnimelist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Refreshtoken",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "SetAnime",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserAnime"
								},
								"examples": {
									"Example 1": {
										"value": {
											"animeId": "animeid not be check of exist!",
											"status": "viewed",
											"review": "awesome!",
											"star": "1-10"
										}
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "animeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UserAnimeStatuses"
						}
					},
					{
						"in": "query",
						"name": "review",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "star",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UserAnimeStars"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Refreshtoken",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteAnime",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "animeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Refreshtoken",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UserSchema"
										},
										{
											"$ref": "#/components/schemas/ErrorSchema"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"email": "test@gmail.com",
											"password": "pas34Wo"
										}
									}
								}
							}
						}
					}
				},
				"description": "User registration by email and password",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "password",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth": {
			"post": {
				"operationId": "Signin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"token": {
													"type": "string"
												},
												"user": {
													"$ref": "#/components/schemas/UserSchema"
												}
											},
											"required": [
												"token",
												"user"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ErrorSchema"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"email": "test@gmail.com",
											"password": "pas34Wo"
										}
									}
								}
							}
						}
					}
				},
				"description": "Authorization of the user by email and password",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "password",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/me": {
			"post": {
				"operationId": "Me",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UserSchema"
										},
										{
											"$ref": "#/components/schemas/ErrorSchema"
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Refreshtoken",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "Refreshtoken",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ongoing": {
			"get": {
				"operationId": "GetOngoing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnimeList"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/search": {
			"get": {
				"operationId": "SearchByQuery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnimeList"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "queryText",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/update": {
			"put": {
				"operationId": "UserUpdate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserSchema"
								},
								"examples": {
									"Example 1": {
										"value": {
											"displayName": "name",
											"photoURL": "https://some.jpg"
										}
									}
								}
							}
						}
					}
				},
				"description": "User update fields",
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Refreshtoken",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "displayName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "photoURL",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}